+++ Mythic_CLI/Changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.27] - 2026=-8-05
+
+### Changed
+
+- Added `jupyter_include` as an option to _not_ include the jupyter notebook docker image in the docker-compose.yml file
+- Made the `docker-compose.yml` write atomic friendly with a temp write + replace
+
 ## [0.3.26] - 2026-03-16
 
 ### Changed